+++ quackosm/__init__.py
@@ -22,7 +22,7 @@ from quackosm.geocode import geocode_to_geometry
 from quackosm.pbf_file_reader import PbfFileReader
 
 __app_name__ = "QuackOSM"
-__version__ = "0.18.1"
+__version__ = "0.18.2"
 
 __all__ = [
     "PbfFileReader",